+++ src/niteoweb/click2sell/interfaces.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
 """Where all interfaces, events and exceptions live."""
 
-from niteoweb.click2sell import Click2SellMessageFactory
+from niteoweb.click2sell import Click2SellMessageFactory as _
 from zope import schema
 from zope.interface import Attribute
 from zope.interface import implements
 from zope.interface import Interface
-
-_ = Click2SellMessageFactory
+from niteoweb.click2sell import parse_mapping
 
 
 # control panel schema
@@ -25,6 +24,19 @@ class IClick2SellSettings(Interface):
             default=u"Enter the Secret Key you got from Click2Sell to access " \
                      "their API."),
         required=True,
+    )
+
+    mapping = schema.List(
+        title=_(u"Product ID to Group mapping"),
+        description=_(u"help_secretkey",
+            default=u"Optionally, you can set Product ID to Group mapping. " \
+            "This is used to automatically add a new member to a certain " \
+            "group based on which Click2Sell product was purchased. Format: " \
+            "'PRODUCT_ID|GROUP_ID'. Example: '1|premium_members'. One " \
+            "mapping per line."),
+        required=False,
+        value_type=schema.ASCIILine(),
+        constraint=parse_mapping,
     )